fix crash when loading Estrangelo Edessa Regular font
git-svn-id: svn://scribus.net/trunk/Scribus@16575 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/fonts/scface_ttf.cpp
@@ -207,7 +207,7 @@ void KernFeature::makeCoverage()
 					}
 					else
 					{
-						for ( unsigned int gl ( Start ); gl >= End; --gl )
+						for ( int gl ( Start ); gl >= (int) End; --gl )
 							coverages[SubTable]  << gl;
 					}
 				}
